Fixed Privacy Policy and footer flex
I wrapped the footer components in a new wrapper labeled "FooterInner", I added similar styles to the Headers version of its wrapper

I added position:relative to the copyright container class and gave it new minimum width breakpoints that help adjust it need be.

I also to get npm start to work, had to delete the lock.json and reinstall on my device.